feat: decouple components in person page
